// src/app/interfaces/property/dto-property.ts
export interface DtoProperty {
    id: number;
    name: string;
    area: number;
    price: string;
    address: string;
    latitude: number;
    longitude: number;
    status: boolean;
    description: string;
    agent: number;
    concept: string;
    type: string;
    images: string[];
    amenities: string[];
}